requirements.txt
seaborn
tensorflow
tqdm
scikit-learn
pillow
